requirements.txt
aws-cdk-lib==2.111.0
constructs>=10.0.0,<11.0.0
boto3==1.33.4
aws-cdk.lambda-layer-kubectl-v28==2.1.0